final
This is a Firebase backed train schedule application. It uses HTML, CSS and jQuery to render the schedule and admin panel on the web page.

Uses Firebase to host a globally available set of data that will be retrieved and manipulated using Momen.js to provide up-to-date information about various trains arrival times and minutes until arrival.